wip making common compile
