+++ src/commonMain/kotlin/core/commands/CommandParsers.kt
@@ -4,6 +4,7 @@ import core.DependencyInjector
 import core.GameState
 import core.Player
 import core.utility.NameSearchableList
+import core.utility.toSortedMap
 import magic.castSpell.CastCommand
 
 object CommandParsers {
