+++ src/commonMain/kotlin/core/properties/Values.kt
@@ -1,9 +1,7 @@
 package core.properties
 
 import core.properties.GuardManager.getGuardedValue
-import core.utility.apply
-import core.utility.hasAllOf
-import core.utility.matches
+import core.utility.*
 
 @kotlinx.serialization.Serializable
 data class Values(private val properties: MutableMap<String, String> = mutableMapOf()) {
